set servers current index, move serverCommands to own file
