+++ api/servers.js
@@ -389,89 +389,13 @@ export function getAllServers(){
 
     const query = "SELECT * FROM servers ORDER BY created ASC";
 
-    return sqliteGetAll(query);
-}
+    const result = sqliteGetAll(query);
 
-export function createServerString(id, server){
+    //set the current .q ids at this time
+    for(let i = 0; i < result.length; i++){
 
-    const idLength = 2;
-    const aliasLength = 25;
-    const mapLength = 25;
-    const playersLength = 7;
-
-    const now = Math.floor(Date.now() * 0.001);
-    //const diff = now - server.modified;
-
-    const serverId = forceStringLength(id, idLength);
-    const alias = forceStringLength(server.alias, aliasLength);
-    
-    let playerString = "";
-
-    if(server.max_players == "ers"){
-        playerString = "Players";
-    }else{
-        playerString = server.players+"/"+server.max_players;
+        result[i].current_index = i + 1;
     }
 
-    /*if(diff >= serverInfoPingInterval * 2 && server.modified !== undefined){
-        server.map = "Timed Out!";
-        playerString = "N/A";
-    }*/
-
-    const map = forceStringLength(server.map, mapLength)+" ";
-    
-    const players = forceStringLength(playerString, playersLength, true);
-
-    return `\`${serverId} - ${alias} ${map} ${players}\``;
-
-}
-
-export function sendNoServers(message, bOnlyActive){
-
-    const title =  "Unreal Tournament Server List";
-
-    const desc = (bOnlyActive) ? `There are currently no active servers.` : "There aren't any servers added to the bot.";
-
-    const embed = new EmbedBuilder()
-    .setColor(embedColor)
-    .setTitle(`${(bOnlyActive) ? "Active" : ""} ${title}`)
-    .setDescription(desc)
-    .setTimestamp();
-
-    return message.channel.send({"embeds": [embed]});
-}
-
-
-export function createServerListParts(servers){
-
-    const parts = [];
-
-    let desc = createServerString("ID", {"alias": "Alias", "map": "Map", "players": 0, "max_players": "ers"}, );
-    desc += `\n`;
-
-    let currentCount = 0;
-
-    for(let i = 0; i < servers.length; i++){
-
-        const s = servers[i];
-
-        desc += createServerString(i + 1, s);
-        if(i < servers.length - 1) desc += `\n`;
-
-        currentCount++;
-
-        if(currentCount >= maxServersPerBlock){
-            currentCount = 0;
-            parts.push(desc);
-            desc = ``;
-        }
-    }
-
-    if(parts.length === 0){
-        parts.push(desc);
-    }else if(desc !== ""){
-        parts.push(desc);
-    }
-
-    return parts;
+    return result;
 }
